add mobile view for operation hours page
