+++ frontend/src/pages/OperationHours.tsx
@@ -1,6 +1,6 @@
 import React, { useEffect, useState } from 'react';
 import { fetchAllLocationOperatingTimes } from '../util/data';
-import { getWeekday } from '../util/helper';
+import { getWeekday, formatTime } from '../util/helper';
 import { OperationHoursData, Hour } from '../types/OperationTypes';
 import { Table, TableBody, TableCell, TableHead, TableHeader, TableRow } from '../components/table'
 
@@ -17,21 +17,11 @@ const OperationHours: React.FC = () => {
     "Chicago Campus": ["Harry's Cafe", "SLICE Pizzeria", "Starbucks (Chicago Campus)"],
   }
 
-
-  const formatTime = (hour: string, minutes: string): string => {
-    const hourInt = parseInt(hour, 10);
-    const minutesInt = parseInt(minutes, 10);
-
-    // Determine AM/PM
-    const period = hourInt >= 12 ? "p" : "a";
-    const formattedHour = hourInt > 12 ? hourInt - 12 : hourInt === 0 ? 12 : hourInt;
-
-    // Format the time as "hh:mm"
-    const formattedTime = `${formattedHour}:${minutesInt.toString().padStart(2, "0")}${period}`;
-
-    return formattedTime;
-  };
-
+  useEffect(() => {
+    fetchAllLocationOperatingTimes().then((data) => {
+      setLoctionOperatingTimes(data.locationOperatingTimes);
+    });
+  }, []);
   const formatHours = (hours: Hour[] | null) => {
     if (hours === null) return "Closed";
 
@@ -43,70 +33,115 @@ const OperationHours: React.FC = () => {
     });
 
     // Join all time ranges with a space in between
-    return formattedHours.map((hour, index) => <span key={index}>{hour}<br /></span>);
+    return formattedHours.map((hour, index) => <span key={index}> {hour} < br /> </span>);
   };
-
-  useEffect(() => {
-    fetchAllLocationOperatingTimes().then((data) => {
-      setLoctionOperatingTimes(data.locationOperatingTimes);
-    });
-  }, []);
 
   // TODOD need to make responsive especially for mobile this looks awful :D 
   return (
     <div>
       {Object.keys(locationGrouping).map((group) => (
-        <div key={group} className="overflow-x-auto mb-8">
+        <div key={group} className="mb-8">
           <h2 className="text-2xl font-bold mb-4 text-gray-900 dark:text-white">{group}</h2>
-          <Table grid className="min-w-full table-auto">
-            <TableHead>
-              <TableRow>
-                <TableHeader className="p-2">Location</TableHeader>
-                {locationOperatingTimes?.[0]?.Week?.map((day) => (
-                  <TableHeader key={day.Date} className="p-2">
-                    {getWeekday(parseInt(day.Day))}
-                    <br />
-                    ({day.Date.slice(5)})
-                  </TableHeader>
-                ))}
-              </TableRow>
-            </TableHead>
-            <TableBody>
-              {locationGrouping[group].map((location) => {
-                const locationData = locationOperatingTimes.find((data) => data.Name === location);
-                return (
-                  <TableRow key={location}>
-                    <TableCell className="p-2">{location}</TableCell>
-                    {locationData?.Week.map((day) => {
-                      const hours = day?.Hours
-                      return (
+
+          {/* Table layout for large screens */}
+          <div className="hidden sm:block overflow-x-auto">
+            <Table grid className="min-w-full table-auto">
+              <TableHead>
+                <TableRow>
+                  <TableHeader className="p-2">Location</TableHeader>
+                  {locationOperatingTimes?.[0]?.Week?.map((day) => (
+                    <TableHeader key={day.Date} className="p-2">
+                      {getWeekday(parseInt(day.Day))}
+                      <br />
+                      ({day.Date.slice(5)})
+                    </TableHeader>
+                  ))}
+                </TableRow>
+              </TableHead>
+              <TableBody>
+                {locationGrouping[group].map((location) => {
+                  const locationData = locationOperatingTimes.find((data) => data.Name === location);
+                  return (
+                    <TableRow key={location}>
+                      <TableCell className="p-2">{location}</TableCell>
+                      {locationData?.Week.map((day) => (
                         <TableCell key={day.Date} className="p-2">
-                          {formatHours(hours)}
+                          {day.Status == "closed" ? "Closed" : formatHours(day?.Hours)}
                         </TableCell>
-                      );
-                    })}
-                  </TableRow>
-                );
-              })}
-            </TableBody>
-          </Table>
+                      ))}
+                    </TableRow>
+                  );
+                })}
+              </TableBody>
+            </Table>
+          </div>
+
+          {/* Card layout for small screens */}
+          <div className="block sm:hidden space-y-4">
+            {locationGrouping[group].map((location) => {
+              const locationData = locationOperatingTimes.find((data) => data.Name === location);
+              return (
+                <div
+                  key={location}
+                  className="p-4 border border-gray-300 rounded-lg shadow-sm bg-white dark:bg-gray-800"
+                >
+                  <h3 className="text-lg font-semibold text-gray-900 dark:text-white mb-2">{location}</h3>
+                  {locationData?.Week.map((day) => (
+                    <div key={day.Date} className="flex justify-between text-sm border-t border-gray-200 py-2">
+                      <span className="text-gray-600 dark:text-gray-400">{getWeekday(parseInt(day.Day))}</span>
+                      <span className="text-gray-800 dark:text-gray-200">{formatHours(day?.Hours)}</span>
+                    </div>
+                  ))}
+                </div>
+              );
+            })}
+          </div>
         </div>
       ))}
     </div>
   );
-};
 
-// <TableBody>
-// {locationOperatingTimes.map((location) => (
-//   <TableRow key={location.Name}>
-//     <TableCell className="font-medium p-2">{location.Name}</TableCell>
-//       {location.Week.map((day) => (
-//         <TableCell key={day.Date} className="p-2">
-//           {day.Status === "closed" ? "Closed" : formatHours(day.Hours)}
-//           </TableCell>
-//       ))}
-//     </TableRow>
-// ))}
-// </TableBody>
+};
+// return (
+// <div>
+//   {Object.keys(locationGrouping).map((group) => (
+//     <div key={group} className="overflow-x-auto mb-8">
+//       <h2 className="text-2xl font-bold mb-4 text-gray-900 dark:text-white">{group}</h2>
+//         <Table grid className="min-w-full table-auto">
+//           <TableHead>
+//             <TableRow>
+//               <TableHeader className="p-2">Location</TableHeader>
+//               {locationOperatingTimes?.[0]?.Week?.map((day) => (
+//                 <TableHeader key={day.Date} className="p-2">
+//                   {getWeekday(parseInt(day.Day))}
+//                     <br />
+//                   ({day.Date.slice(5)})
+//                 </TableHeader>
+//               ))}
+//             </TableRow>
+//           </TableHead>
+//           <TableBody>
+//             {locationGrouping[group].map((location) => {
+//               const locationData = locationOperatingTimes.find((data) => data.Name === location);
+//               return (
+//                 <TableRow key={location}>
+//                   <TableCell className="p-2">{location}</TableCell>
+//                     {locationData?.Week.map((day) => {
+//                       const hours = day?.Hours
+//                       return (
+//                         <TableCell key={day.Date} className="p-2">
+//                           {formatHours(hours)}
+//                           </TableCell>
+//                       );
+//                     })}
+//                   </TableRow>
+//               );
+//             })}
+//           </TableBody>
+//         </Table>
+//       </div>
+//   ))}
+//   </div>
+// );
 export default OperationHours;
